create client-side auth unit tests; add dofetch unit and integration tests
