+++ client/common/test/dofetch.unit.spec.js
@@ -1,5 +1,5 @@
 import tape from 'tape'
-import { processFormData } from '../dofetch.js'
+import * as df from '../dofetch.js'
 
 /*************************
  reusable helper functions
@@ -11,13 +11,15 @@ const text2buf = new TextEncoder()
  test sections
 ***************/
 
+console.log(`-***- common/dofetch unit -***-`)
+
 tape('processFormData', async test => {
 	const form = new FormData()
 	const blob = new Blob(text2buf.encode('Hello, wold'), { type: 'application/octet-stream' })
 	form.append('gzfile', blob, 'myfile')
 	form.append('errors', '')
 	try {
-		const parts = await processFormData({
+		const parts = await df.processFormData({
 			formData() {
 				return form
 			}
@@ -28,5 +30,18 @@ tape('processFormData', async test => {
 	} catch (e) {
 		console.log(e)
 	}
+	test.end()
+})
+
+tape('setAuth()', async test => {
+	const opts = {
+		dsAuth: [
+			{ dslabel: 'abc', type: 'basic', insession: false },
+			{ dslabel: 'xyz', type: 'basic', insession: true }
+		]
+	}
+	df.setAuth(opts, df.dofetch3)
+	test.equal(df.isInSession('abc'), false, 'should detect a dslabel that is not in session')
+	test.equal(df.isInSession('xyz'), true, 'should detect a dslabel that is in session')
 	test.end()
 })
